Avoid unlink of upload file in Quercus because there is proper cleanup there regardless if we use move_uploaded_file or not. Unlink failed at early tests of multipart upload, but maybe it doesn't anymore. We might consider to revert this change before release.
git-svn-id: https://optime.repos.se/svn/code/repos/trunk@6175 94f62164-8fde-0310-83d7-e1105c8c1596

+++ repos-web/edit/upload/index.php
@@ -274,7 +274,9 @@ class Upload {
 			}
 			fclose($to);
 			fclose($from);
-			unlink($current); // because we don't use move_uploaded_file anymore
+			if (!isReposJava()) { // removes uploaded tmp automatically
+				unlink($current); // because we don't use move_uploaded_file anymore
+			}
 		} else {
 			if ($n = $this->getOriginalFilename()) trigger_error("Could not access the uploaded file ", E_USER_ERROR);
 			Validation::error('A local file must be selected for upload');
